Initial commit
This image is based on Alpine Linux. It follows the practices followed in the
github.com/docker-library organization. It could become part of the "Official
Images" programs if it's ever submitted.

It only supports to queue types: `builtin` and `redis`. `libmemcached` was
intended to be added but I couldn't get it working in Alpine Linux. I think the
problem is in the way that the libmemcached-dev is built with `static`
disabled.