Set secure SHELL options before RUN
Docker only evaluates the exit code of the last operation after the pipe
in a RUN statement to determine success. Specifying 'pipefail' will cause
the build to fail due to an error at any stage of a RUN command with
pipes.

Addresses Hadolint DL4006 warning: Set the SHELL option -o pipefail
before RUN with a pipe in it.

Setting the -e option instructs bash to immediately exit if any command
in a RUN statement has a non-zero exit status.

Setting the -u option triggers an error if a variable is referenced that
has not been previously defined.

Setting runDeps using a list passed to 'apk info --installed' returns a
non zero exit status if one of the packages listed is not installed.
Added '|| true' to force an exit status of 0.